Rag_based_Chart_bot_with_local_llama_tinymodel-main/requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
supabase
httpx
pydantic
PyPDF2
requests